==> 02_Referrals_Cleaning.ipynb <==
# Partners for Healing Referrals

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import plotly.express as px
import plotly.graph_objs as go
from plotly.offline import init_notebook_mode, iplot

init_notebook_mode(connected=True)

## Importing and Cleaning Dataframe

#importing the dataframe
re = pd.read_csv('referrals.csv')

#renaming the columns
re.rename(columns = re.loc[0], inplace=True)
re.drop(0, axis = 0, inplace=True)

# function that returns the year from a date format of 'mm/dd/year'
def dayToYear(day):
    return day.split('/')[-1]

# creating the new column
re['enc year'] = [dayToYear(day) for day in re['order chartdate']]

re['enc year']=re['enc year'].astype(int)

re = re.drop(['patientdob','order name (single)','order desc'], axis=1)

re.to_csv('Referral_Data.csv')
