==> 03_Hear_About_Cleaning.ipynb <==
# Hear About Partners for Healing

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import plotly.express as px
import plotly.graph_objs as go
from plotly.offline import init_notebook_mode, iplot

init_notebook_mode(connected=True)

## Importing and Cleaning Dataframe

#import the csv
ha = pd.read_csv('HearPfH.csv')

#change the column names and drop a column
ha.rename(columns = ha.loc[0], inplace=True)
ha.drop(0, axis = 0, inplace=True)
ha = ha.drop('patient name', axis=1)

#simplify the 'ref column' outputs
ha['ref source'] = ha['ref source'].replace('Other : Cigna Connect INS', 'Other')
ha['ref source'] = ha['ref source'].replace('Other : over 1 yr since last visit', 'Other')
ha['ref source'] = ha['ref source'].replace('Referral from Social Service Agency', 'Social Service Agency')
ha['ref source'] = ha['ref source'].replace('Referral from Other Clinic/Physician', 'Other Clinic/Physician')
ha['ref source'] = ha['ref source'].replace('Referral from Hospital/ER', 'Hospital/ER')
ha['ref source'] = ha['ref source'].replace('Other : over 1 yr since last visit', 'Other')

ha.to_csv('FinalHearAbout.csv')
