==> 04_State_Census.ipynb <==
import pandas as pd
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import plotly.express as px
import plotly.graph_objs as go
from plotly.offline import init_notebook_mode, iplot

init_notebook_mode(connected=True)  


# Import Original Census CSV Files

s2012 = pd.read_csv('02_ACSST5Y2012.S2701.csv')
s2013 = pd.read_csv('02_ACSST5Y2013.S2701.csv')
s2014 = pd.read_csv('02_ACSST5Y2014.S2701.csv')
s2015 = pd.read_csv('02_ACSST5Y2015.S2701.csv')
s2016 = pd.read_csv('02_ACSST5Y2016.S2701.csv')
s2017 = pd.read_csv('02_ACSST5Y2017.S2701.csv')
s2018 = pd.read_csv('02_ACSST5Y2018.S2701.csv')
s2019 = pd.read_csv('02_ACSST5Y2019.S2701.csv')
s2020 = pd.read_csv('02_ACSST5Y2020.S2701.csv')
s2021 = pd.read_csv('02_ACSST5Y2021.S2701.csv')

us2012 = pd.read_csv('03_ACSST5Y2012.S2701.csv')
us2013 = pd.read_csv('03_ACSST5Y2013.S2701.csv')
us2014 = pd.read_csv('03_ACSST5Y2014.S2701.csv')
us2015 = pd.read_csv('03_ACSST5Y2015.S2701.csv')
us2016 = pd.read_csv('03_ACSST5Y2016.S2701.csv')
us2017 = pd.read_csv('03_ACSST5Y2017.S2701.csv')
us2018 = pd.read_csv('03_ACSST5Y2018.S2701.csv')
us2019 = pd.read_csv('03_ACSST5Y2019.S2701.csv')
us2020 = pd.read_csv('03_ACSST5Y2020.S2701.csv')
us2021 = pd.read_csv('03_ACSST5Y2021.S2701.csv')

# Create and Clean the Census Dataframes

# create a list of tn census df and list of years
tn_dataframes = [s2012, s2013, s2014, s2015, s2016, s2017, s2018, s2019, s2020, s2021]
tn_years = []
for i in range(2012, 2022):
    tn_years.append(i)

# create a list of us census df and list of years    
us_dataframes = [us2012, us2013, us2014, us2015, us2016, us2017, us2018, us2019, us2020, us2021]
us_years = []
for i in range(2012, 2022):
    us_years.append(i)

#rename the columns of the df
def rename_columns_smaller(df, year):
    if df.shape[1]==4:
        df.columns=['del', 'population', 'uninsured', 'percent_uninsured']
    else:
        df.columns = ['del', 'population', 'insured', 'percent_insured', 'uninsured', 'percent_uninsured']
        df = df.drop(['insured','percent_insured'], axis=1)
    df = df.head(1)
    df = df.drop('del', axis=1)
    df['year'] = year  # Add year column
    return df

#modify the tn df
tn_modified_dataframes = []
for i, df in enumerate(tn_dataframes):
    tn_modified_df = rename_columns_smaller(df, tn_years[i])
    tn_modified_dataframes.append(tn_modified_df)


#modify the us df
us_modified_dataframes = []
for i, df in enumerate(us_dataframes):
    us_modified_df = rename_columns_smaller(df, us_years[i])
    us_modified_dataframes.append(us_modified_df)

# concat the list of tn df
tn_df = pd.concat(tn_modified_dataframes, ignore_index=True)

# concat the list of us df
us_df = pd.concat(us_modified_dataframes, ignore_index=True)

# cleaning the data type for df
def type_cleaning(df):
    #Making 'percent uninsured' column float types
    df['percent_uninsured'] = df['percent_uninsured'].str.replace('%', '')
    df['percent_uninsured'] = df['percent_uninsured'].astype(float) 

    df['percent_uninsured'] = pd.to_numeric(df['percent_uninsured'], errors='coerce')

# using the cleaning func
type_cleaning(tn_df)

#using the cleaning func
type_cleaning(us_df)

# correct percetnage values
def subtract_from_100(value):
    if value > 50:
        return 100 - value
    else:
        return value

# correct the values
def correct_percent(df):
    df['percent_uninsured'] = df['percent_uninsured'].apply(subtract_from_100)

# final df
correct_percent(tn_df)
correct_percent(us_df)

# Download the final csv files
tn_df.to_csv('tn_trends.csv')
us_df.to_csv('us_trends.csv')
