==> 06_Zip_Code_Census_Cleaning.ipynb <==
import pandas as pd

#import original census data by zip codes
big = pd.read_csv('ACSST5Y2020.S2701-Data.csv')

# Clean the Census Dataframe

#Change the column names to the first row of data
big.rename(columns = big.loc[0], inplace=True)
big.drop(0, axis = 0, inplace=True)

#Changing the zip code values to only include the zip code and not 'ZCTA5'
big['Geographic Area Name'] = big['Geographic Area Name'].str.replace('ZCTA5 ', '')

#Creating a smaller dataframe that only includes total population, insured population, and uninsured population
small = big[['Geographic Area Name','Estimate!!Total!!Civilian noninstitutionalized population','Estimate!!Insured!!Civilian noninstitutionalized population','Estimate!!Uninsured!!Civilian noninstitutionalized population']]

#Renaming the new dataframe's columns
small.columns = ['zip code','population','insured','uninsured']

#Changing the columns to numeric types
small['population'] = pd.to_numeric(small['population'])
small['insured'] = pd.to_numeric(small['insured'])
small['uninsured'] = pd.to_numeric(small['uninsured'])

#Creating two percentage columns for insured and uninsured
small = small.assign(percent_insured=small['insured'] / small['population']*100)
small = small.assign(percent_uninsured=small['uninsured'] / small['population']*100)

#Rounding the columns to 2 decimal places
small['percent_insured']=small['percent_insured'].round(2)
small['percent_uninsured']=small['percent_uninsured'].round(2)

#Creating a tennessee zip code list using the small dataframe
zips = small['zip code'].tolist()

#creating new df
df_zip = small.drop(['population',
 'insured',
 'uninsured',
 'percent_insured',
 'percent_uninsured'], axis=1)

# Download the Clean Census Dataframe

#Create census df
small.to_csv('zipcensusdata.csv')

# Import and Clean the uszips.csv

#import uszips from https://simplemaps.com/data/us-zips
df = pd.read_csv('uszips.csv')

#sort the dataframe to only include TN values
df = df[df['state_id']=='TN']

# drop columns that are not important
df = df.drop(['lat','lng',
 'zcta',
 'parent_zcta',
 'population',
 'density',
 'county_fips',
 'county_weights',
 'county_fips_all',
 'imprecise',
 'military',
 'timezone'], axis=1)

# Download the Clean Zip Code Dataframe

df.to_csv('tnzips.csv')
