==> 07_Shape_Cleaning.ipynb <==
import json
import pandas as pd
import plotly.express as px
import geopandas as gpd

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import plotly.graph_objs as go
from plotly.offline import init_notebook_mode, iplot

init_notebook_mode(connected=True)  

# County Shape File Creation

#import shape file
og_tn_shape = gpd.read_file('cb_2022_us_county_500k.shp')
#make a copy of shape file
tn_shape = og_tn_shape.copy(deep = True)
#remove geometry
geocol = tn_shape.pop('geometry')
#add geometry to the first column
tn_shape.insert(0, 'geometry', geocol)
#filter the state codes
tn_shape = tn_shape[tn_shape['STUSPS'] == 'TN']
#rename the county column
tn_shape.rename(columns={'NAMELSAD': 'county'}, inplace=True)
#drop none value row
tn_shape.dropna(axis = 0, subset = 'geometry', how = 'any', inplace = True)
#import dataframe
df = pd.read_csv('CountyCensus2021.csv')
#filter dataframe years 
df = df[df['year']==2021]


#merging the shape and dataframe
tn_shape = tn_shape.merge(df, on='county')

tn_shape.columns = ['geometry', 'STATEFP', 'COUNTYFP', 'COUNTYNS', 'AFFGEOID', 'GEOID',
       'NAME', 'county', 'STUSPS', 'STATE_NAME', 'LSAD', 'ALAND', 'AWATER',
       'population', 'percent_uninsured', 'uninsured', 'percent_insured',
       'insured', 'year']

#save the new shape file
tn_shape.to_file('filtered_tn_shapefile.shp')

# Zip Shape File Creation

#import shape file
og_zip_shape = gpd.read_file('tl_2022_us_zcta520.shp')
#make a copy of shape file
zip_shape = og_zip_shape.copy(deep = True)
#remove geometry
geocol = zip_shape.pop('geometry')
#add geometry to the first column
zip_shape.insert(0, 'geometry', geocol)
#rename the zip column
zip_shape.rename(columns={'ZCTA5CE20':'zip'}, inplace=True)
#drop none value row
zip_shape.dropna(axis = 0, subset = 'geometry', how = 'any', inplace = True)
#import tn census data
tc = pd.read_csv('tnzips.csv')
#drop the unnamed column
tc.drop(['Unnamed: 0'], axis = 1, inplace=True)
#import tn zips from 
df = pd.read_csv('zipcensusdata.csv')
#drop the unnamed column
df.drop(['Unnamed: 0'], axis = 1, inplace=True)
#rename the zip code column
df.rename(columns={'zip code':'zip'}, inplace=True)
#merge the census and zip code dataframes
df = df.merge(tc, on='zip')
#filter specific counties
df = df[df['county_name'].isin(['Coffee','Franklin', 'Moore', 'Marion', 'Grundy', 'Warren', 'Cannon', 'Bedford'])]
#change zip column to str
df['zip']=df['zip'].astype(str)
#merge shape and dataframe
zip_shape = zip_shape.merge(df, on='zip')

#save the new shapefile
zip_shape.to_file('filtered_zip_shapefile.shp')
