==> 08_Population_Reached_DF.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import plotly.graph_objs as go
from plotly.offline import init_notebook_mode, iplot

init_notebook_mode(connected=True)  

# Population Reached Dataframe

def population_reached():    
    gen_census = pd.read_csv('01general_census.csv')
    cond = pd.read_csv('Encounter_Conditions.csv')
    
    area = ['Coffee County', 'Franklin County', 'Moore County']
    pfh_area = ['Coffee', 'Franklin', 'Moore']
    years4 = [2018, 2019, 2020, 2021]

    county3 = gen_census[gen_census['county'].isin(area)]

    cond_county = cond[cond['ptnt cnty f rsdnc'].isin(pfh_area)]

    cond_county = cond_county.drop(['index', 'patientdob', 'patientsex', 'patientregd',
           'status', 'ptnt rsn fr nctv stts', 'ptnt dcsd ysn', 'ptnt  fpl',
           'patient city', 'patient zip', 'race', 'ethnicity',
           'patient lang', 'cln enc date', 'appttype', 'prvdr', 'auth refto prvdr',
           'icd10encounterdiagdescr', 'icd9encounterdiagdescr',
           'patientsnomedproblemcodedesc', 'enc srv diag', 'age',
           'enc month', 'enc day', 'distance'], axis=1)

    first_df = cond_county.groupby(['enc year', 'patientid']).first().reset_index()

    first_df = first_df[first_df['enc year'].isin(years4)]

    grouped_df = first_df.groupby(['ptnt cnty f rsdnc', 'enc year', 'patientid']).size().reset_index(name='count')


    result = grouped_df.groupby(['ptnt cnty f rsdnc', 'enc year']).size().reset_index(name='patient_count')

    result.columns = ['county','year','patient_count']

    result['county'] = result['county'] + ' County'

    merged_df = pd.merge(result, county3, on=['county', 'year'])

    merged_df = merged_df.drop(['Unnamed: 0'], axis=1)

    merged_df['percent_reached'] = merged_df['patient_count'] / merged_df['uninsured']*100

    pop_reached = merged_df
    
    return pop_reached

pop_reached = population_reached()

#pop_reached.to_csv('Population_Reached.csv')
